Fixed CLIs; Refactored code.
