+++ converters/mbtiles/cli
@@ -1,24 +1,28 @@
 #!/usr/bin/env node
 
-var MBTilesToGeoPackage = require('./index');
-var path = require('path');
+const MBTilesToGeoPackage = require('./built/index');
+const path = require('path');
 
 console.log('MBTiles to GeoPackage Converter');
 
-var mbtilesFile = path.resolve(process.cwd(), process.argv[2]);
-var geoPackageFile = path.resolve(process.cwd(), process.argv[3]);
+const mbtilesFile = path.resolve(process.cwd(), process.argv[2]);
+const geoPackageFile = path.resolve(process.cwd(), process.argv[3]);
 
 console.log('Converting %s to %s', mbtilesFile, geoPackageFile);
-MBTilesToGeoPackage.convert({
-  mbtiles: mbtilesFile,
-  geopackage: geoPackageFile
-}, function(status, callback) {
-  callback();
-}, function(err, result) {
-  if (err) {
-    console.log('File conversion failed with error: ', err);
-  } else {
-    console.log('File conversion complete');
-    console.log('Converted %s to %s', mbtilesFile, geoPackageFile);
-  }
-});
+MBTilesToGeoPackage.convert(
+  {
+    mbtiles: mbtilesFile,
+    geopackage: geoPackageFile,
+  },
+  function(status, callback) {
+    callback();
+  },
+  function(err, result) {
+    if (err) {
+      console.log('File conversion failed with error: ', err);
+    } else {
+      console.log('File conversion complete');
+      console.log('Converted %s to %s', mbtilesFile, geoPackageFile);
+    }
+  },
+);
